+++ src/components/ChatWindow/ChatWindowBox.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import PropType from 'prop-types';
-import ChatWindow from './ChatWindow';
 import theme from '../../themes/_theme_two.scss';
-import themeb from '../../themes/_theme_three.scss';
+// import theme from '../../themes/_theme_three.scss';
+import ChatWindowProvider from './ChatWindowProvider';
 
 class ChatWindowBox extends Component {
   constructor(props) {
@@ -11,7 +11,6 @@ class ChatWindowBox extends Component {
     this.state = {
       messages: [],
       currentUserId: userId,
-      self: true,
       other: true,
     };
   }
@@ -21,7 +20,6 @@ class ChatWindowBox extends Component {
     if (currentUserId !== nextProps.userId) {
       this.setState({
         messages: [],
-        self: true,
         other: true,
       });
     }
@@ -37,7 +35,6 @@ class ChatWindowBox extends Component {
   handleChatBoxClose = (win) => {
     if (win === '1') {
       this.setState({
-        self: false,
         other: false,
       });
     } else {
@@ -48,33 +45,19 @@ class ChatWindowBox extends Component {
   }
 
   render() {
-    const { self, messages, other } = this.state;
+    const { messages, other } = this.state;
     const { myData, userData, userId } = this.props;
     return (
       <div className="demo-chat-window-box">
-        {self && (
-        <div className="demo-chat-window self">
-          <ChatWindow
-            theme={theme}
-            handleMessages={this.handleMessages}
-            myData={myData}
-            userData={userData.find(user => user.id === userId)}
-            messages={messages}
-            handleCloseClick={() => this.handleChatBoxClose('1')}
-          />
-        </div>
-        )}
         {other && (
-        <div className="demo-chat-window other">
-          <ChatWindow
-            theme={themeb}
+          <ChatWindowProvider
+            theme={theme}
             handleMessages={this.handleMessages}
             userData={myData}
             myData={userData.find(user => user.id === userId)}
             messages={messages}
             handleCloseClick={() => this.handleChatBoxClose('2')}
           />
-        </div>
         )}
       </div>
     );
