+++ src/components/index.js
@@ -3,8 +3,7 @@ export { default as ChatListItem } from './ChatList/ChatListItem';
 export { default as ChatListHeader } from './ChatList/ChatListHeader';
 export { default as ChatList } from './ChatList/ChatList';
 export { default as ChatListSearch } from './ChatList/ChatListSearch';
-export { default as ChatProvider } from './ChatList/ChatProvider';
-export { default as ChatWindow } from './ChatWindow/ChatWindow';
+export { default as ChatProvider } from './ChatList/ChatListProvider';
 export { default as ChatWindowBox } from './ChatWindow/ChatWindowBox';
 export { default as ChatWindowHeader } from './ChatWindow/ChatWindowHeader';
 export { default as ChatWindowSend } from './ChatWindow/ChatWindowSend';
