+++ src/themes/_default_theme.scss
@@ -1,16 +1,9 @@
 @import '../root/colors.scss';
+// Chat list styles.
 :local(.provider) {
     background: $chat-bg;
     box-shadow: 2px 3px 7px 2px rgba(0, 0, 0, 0.5);
-    width: 300px;
-}
-
-:local(.userlist) {
-    margin: 5px 0;
-    padding: 0 20px;
-    overflow-y: scroll;
-    height: 60vh;
-    color: $text-color;
+    width: 220px;
 }
 
 :local(.header) {
@@ -19,32 +12,15 @@
     background: $chat-header;
     color: $chat-head-text;
     padding: 10px 10px;
+    font-size: 0.84em;
 }
 
-:local(.windowHeader) {
-    background: $chat-header;
-    padding: 5px 10px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    color: $chat-head-text;
-    border-radius: 5px 5px 0 0;
-}
-
-:local(.chatWindowAvatar) {
-    border-radius: 50%;
-    max-height: 40px;
-    max-width: 40px;
-    margin-right: 20px;
-}
-
-:local(.chatUserName) {
-    flex: 2;
-    text-transform: capitalize;
-}
-
-:local(.icon) {
-    padding: 9%;
+:local(.userlist) {
+    margin: 5px 0;
+    padding: 0 20px;
+    overflow-y: scroll;
+    height: 40vh;
+    color: $text-color;
 }
 
 :local(.useritem) {
@@ -53,10 +29,13 @@
     justify-content: space-around;
     margin: 10px 0;
     cursor: pointer;
+    font-size: 0.9em;
 }
 
 :local(.useravatar) {
     border-radius: 50%;
+    max-height: 40px;
+    width: auto;
 }
 
 :local(.username) {
@@ -74,8 +53,8 @@
 }
 
 :local(.offline) {
-    height: 15px;
-    width: 20px;
+    font-size: 0.7em;
+    color: #a0a0a0;
 }
 
 :local(.searchbox) {
@@ -107,8 +86,47 @@
     box-shadow: 2px 4px 9px 3px rgba(0, 0, 0, 0.5);
 }
 
+// Chat window styles.
+:local(.chatwindowbox) {
+    display: flex;
+    width: 50vw;
+    position: fixed;
+    left: 0.5em;
+    bottom: 0.5em;
+    z-index: 1000;
+    justify-content: space-between;
+}
+
+:local(.windowHeader) {
+    background: $chat-header;
+    padding: 5px 10px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    color: $chat-head-text;
+    border-radius: 5px 5px 0 0;
+}
+
+:local(.chatWindowAvatar) {
+    border-radius: 50%;
+    max-height: 35px;
+    max-width: 35px;
+    margin-right: 20px;
+}
+
+:local(.chatUserName) {
+    flex: 2;
+    text-transform: capitalize;
+}
+
+:local(.icon) {
+    padding: 9%;
+}
+
 :local(.chatWindowBody) {
-    height: 50vh;
+    height: 40vh;
+    width: 220px;
+    background: 'blue';
     padding: 5px 10px;
     overflow-y: scroll;
     .animated {
@@ -140,6 +158,7 @@
     background: #b8dff0;
     float: right;
 }
+
 :local(.recieved) {
     background: #b8dff0;
     float: left;
@@ -171,7 +190,7 @@
     }
 }
 
-:local(.messageInput){
+:local(.messageInput) {
     font-size: 0.85em;
     border: none;
     outline: none;
