+++ src/themes/_theme_three.scss
@@ -18,7 +18,7 @@
 }
 
 :local(.provider) {
-    width: 300px;
+    width: 220px;
     background: $chat-bg;
     box-shadow: 2px 3px 7px 2px rgba(0, 0, 0, 0.5);
     border-radius: 5px 5px 0 0;
@@ -111,7 +111,7 @@
 }
 
 :local(.chatWindowBody) {
-    height: 50vh;
+    height: 40vh;
     padding: 5px 10px;
     overflow-y: scroll;
     background-color: #ffffff;
